+++ src/ext/java/org/opentripplanner/ext/legacygraphqlapi/datafetchers/LegacyGraphQLPlaceInterfaceTypeResolver.java
@@ -7,7 +7,7 @@ import graphql.schema.TypeResolver;
 import org.opentripplanner.model.Stop;
 import org.opentripplanner.routing.graphfinder.PatternAtStop;
 import org.opentripplanner.routing.bike_park.BikePark;
-import org.opentripplanner.routing.bike_rental.BikeRentalStation;
+import org.opentripplanner.routing.vehicle_rental.VehicleRentalStation;
 
 public class LegacyGraphQLPlaceInterfaceTypeResolver implements TypeResolver {
 
@@ -17,7 +17,7 @@ public class LegacyGraphQLPlaceInterfaceTypeResolver implements TypeResolver {
     GraphQLSchema schema = environment.getSchema();
 
     if (o instanceof BikePark) { return schema.getObjectType("BikePark"); }
-    if (o instanceof BikeRentalStation) { return schema.getObjectType("BikeRentalStation"); }
+    if (o instanceof VehicleRentalStation) { return schema.getObjectType("BikeRentalStation"); }
     // if (o instanceof CarPark) { return schema.getObjectType("CarPark"); }
     if (o instanceof PatternAtStop) { return schema.getObjectType("DepartureRow"); }
     if (o instanceof Stop) { return schema.getObjectType("Stop"); }